TMDB DATASET ANALYSIS
This dataset is from TMDB('the movie database') and contains a various set of features that are being
explored using a set of questions to get insights.

It includes the following sections:

- Data Collection
This involves loading of data into the dataset

- Data Wrangling
This involves cleanig and organising the data for  analyis

- Data Analysis
This section is focused on exploring the data using the given set of questions

- Conclusion
This is a summary of the analysis highlighting key areas.